# Fix number of threads for pushing to views