--- a/tests/queries/0_stateless/01275_parallel_mv.sql.j2
+++ b/tests/queries/0_stateless/01275_parallel_mv.sql.j2
@@ -1,3 +1,6 @@
+-- Tags: no-s3-storage
+-- no-s3-storage: s3 has 20 more threads
+
 -- avoid settings randomization by clickhouse-test
 set max_threads = 0;