… tests/queries/0_stateless/01275_parallel_mv.sql
… tests/queries/0_stateless/02350_views_max_insert_threads.reference