### Changelog category (leave one):
- Bug Fix (user-visible misbehavior in official stable or prestable release)

### Changelog entry (a user-readable short description of the changes that goes to CHANGELOG.md):
Fix `parallel_view_processing=1` with `optimize_trivial_insert_select=1`. Fix `max_insert_threads` while pushing to views.

Refs: #12085
Refs: #29786
Refs: #28582
Fixes: #37900 (cc @filimonov )
Cc: @KochetovNicolai 

_Note: these two fixes in different patches._