@azat 2022-07-03
>[Performance Comparison [4/4] — Failed to parse the report.](https://s3.amazonaws.com/clickhouse-test-reports/38731/b3c0e396ed41cb9aaadd30590f1319e2d71e91f8/performance_comparison_[4/4]/runlog.log)

```
output/clickhouse
tar: clickhouse: Wrote only 3072 of 10240 bytes
2022-07-03 11:14:10 URL:https://s3.amazonaws.com/clickhouse-builds/38731/b3c0e396ed41cb9aaadd30590f1319e2d71e91f8/package_release/performance.tgz [656960872/656960872] -> "-" [1]
output/clickhouse-copier
output/clickhouse-disks
tar: Exiting with failure status due to previous errors
```