@azat 2022-07-03
>[Performance Comparison [3/4] — 2 too long, 5 faster, 1 slower, 1 unstable](https://s3.amazonaws.com/clickhouse-test-reports/38731/7427adb600633008af5688e394eb287f370e1d41/performance_comparison_[3/4]/report.html)

Old, s | New, s | Ratio of speedup (-) or slowdown (+) | Relative difference (new − old) / old | p < 0.01 threshold | Test | # | Query
-- | -- | -- | -- | -- | -- | -- | --
6.894 | 0.825 | -8.356x | -0.881 | 0.880 | views_max_insert_threads | 0 | `insert into views_max_insert_threads_null select * from numbers_mt(3000000000) settings max_threads = 16, max_insert_threads=16`