@azat 2022-07-03
>[Stress test (memory, actions) — Backward compatibility check: Error message in clickhouse-server.log (see bc_check_error_messages.txt)](https://s3.amazonaws.com/clickhouse-test-reports/38731/7427adb600633008af5688e394eb287f370e1d41/stress_test__memory__actions_.html)

```
2022.07.03 18:02:45.335507 [ 388363 ] {} <Error> test_29.tp_1 (07040515-bc40-4602-adf2-83c9d6e37a37): Cannot quickly remove directory /var/lib/clickhouse/store/070/07040515-bc40-4602-adf2-83c9d6e37a37/delete_tmp_all_0_0_0_2 by removing files; fallback to recursive removal. Reason: Code: 458. DB::ErrnoException: Cannot unlink file /var/lib/clickhouse/store/070/07040515-bc40-4602-adf2-83c9d6e37a37/delete_tmp_all_0_0_0_2/pp.proj, errno: 21, strerror: Is a directory. (CANNOT_UNLINK) (version 22.7.1.1)
```

Projections cannot be removed quickly, but this will go away after #38717